requirements.txt
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.0
MarkupSafe==1.1.1
mongoengine==0.19.1
Pillow==6.2.1
pymongo==3.10.1
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
visitor==0.1.3
Werkzeug==0.16.1
WTForms==2.2.1
